requirements.txt
python-telegram-bot==20.7
Pillow
python-dotenv
sqlite3-binary
